[STAGES 1-4] Refactoring
Refactor unit and integration tests
Refactor classes across the application
Change project structure
Update postman collection
Clean up files with db migrations
Add javadoc to the part of the btp and core packages
Add authentication to the Swagger UI

+++ src/test/integration/org/example/learningsystem/course/controller/CourseEnrollmentControllerIT.java
@@ -1,57 +1,40 @@
 package org.example.learningsystem.course.controller;
 
-import org.example.learningsystem.config.DataSourceProperties;
+import org.example.learningsystem.AbstractCommonIT;
 import lombok.RequiredArgsConstructor;
-import org.example.learningsystem.config.PostgreSQLConfiguration;
 import org.example.learningsystem.course.model.Course;
 import org.example.learningsystem.course.service.CourseService;
 import org.example.learningsystem.student.model.Student;
 import org.example.learningsystem.student.service.StudentService;
-import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.test.context.support.WithAnonymousUser;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.servlet.MockMvc;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 import java.math.BigDecimal;
 import java.util.UUID;
 
-import static org.example.learningsystem.util.CourseTestUtils.buildCourse;
-import static org.example.learningsystem.util.StudentTestUtils.buildStudent;
+import static java.util.Objects.nonNull;
+import static org.example.learningsystem.course.common.builder.CourseEnrollmentRequestBuilder.buildEnrollRequest;
+import static org.example.learningsystem.course.common.builder.CourseEnrollmentRequestBuilder.buildUnenrollRequest;
+import static org.example.learningsystem.course.common.builder.CourseBuilderIT.buildCourse;
+import static org.example.learningsystem.student.common.builder.StudentBuilderIT.buildStudent;
+import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@Tag("integration")
-@SpringBootTest
-@AutoConfigureMockMvc
-@ActiveProfiles("test")
-@EnableConfigurationProperties(DataSourceProperties.class)
-@Import(PostgreSQLConfiguration.class)
-@Testcontainers
 @RequiredArgsConstructor
-class CourseEnrollmentControllerIT {
+@WithMockUser
+class CourseEnrollmentControllerIT extends AbstractCommonIT {
 
-    private static final String ENROLLMENT_URL = "/courses/{id}/students/{studentId}";
-
-    @Autowired
-    private MockMvc mockMvc;
     @Autowired
     private CourseService courseService;
     @Autowired
     private StudentService studentService;
 
     @Test
-    @WithMockUser
     void enrollStudent_givenCourseAndStudent_shouldSuccessfullyEnrollStudentAndReturn200() throws Exception {
         // given
         var course = buildCourse();
@@ -61,27 +44,31 @@ class CourseEnrollmentControllerIT {
         var student = buildStudentWithCoins(course.getPrice());
         var savedStudent = studentService.create(student);
         var studentId = savedStudent.getId();
+
         var pageable = PageRequest.of(0, 1);
 
         // when
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isOk());
 
         // then
-        var courseStudents = studentService.getAllByCourseId(courseId, pageable);
-        assertEquals(1, courseStudents.getTotalElements());
-        var updatedStudent = courseStudents.getContent().getFirst();
-        assertEquals(studentId, updatedStudent.getId());
-
-        assertNotEquals(student.getCoins(), updatedStudent.getCoins());
+        var courseStudentsPage = studentService.getAllByCourseId(courseId, pageable);
+        var courseStudents = courseStudentsPage.getContent();
+        var updatedStudent = courseStudents.getFirst();
         var updatedCourse = courseService.getById(courseId);
-        assertNotEquals(course.getCoinsPaid(), updatedCourse.getCoinsPaid());
+
+        assertAll(
+                () -> assertEquals(1, courseStudentsPage.getTotalElements()),
+                () -> assertEquals(studentId, updatedStudent.getId()),
+                () -> assertNotEquals(student.getCoins(), updatedStudent.getCoins()),
+                () -> assertNotEquals(course.getCoinsPaid(), updatedCourse.getCoinsPaid())
+        );
 
         cleanUp(courseId, studentId);
     }
 
     @Test
-    @WithMockUser
     void enrollStudent_givenCourseAndStudent_shouldThrowEnrollmentDeniedExceptionAndReturn500() throws Exception {
         // given
         var course = buildPrivateCourse();
@@ -93,14 +80,14 @@ class CourseEnrollmentControllerIT {
         var studentId = savedStudent.getId();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isInternalServerError());
 
         cleanUp(courseId, studentId);
     }
 
     @Test
-    @WithMockUser
     void enrollStudent_givenCourseAndStudent_shouldThrowInsufficientFundsExceptionAndReturn500() throws Exception {
         // given
         var course = buildPrivateCourse();
@@ -112,57 +99,61 @@ class CourseEnrollmentControllerIT {
         var studentId = savedStudent.getId();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isInternalServerError());
 
         cleanUp(courseId, studentId);
     }
 
     @Test
+    @WithAnonymousUser
     void enrollStudent_givenCourseIdAndStudentId_shouldReturn401() throws Exception {
         // given
         var courseId = UUID.randomUUID();
         var studentId = UUID.randomUUID();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isUnauthorized());
     }
 
     @Test
-    @WithMockUser
     void enrollStudent_givenCourseAndStudentId_shouldThrowEntityNotFoundExceptionAndReturn404() throws Exception {
         // given
         var course = buildPrivateCourse();
         var savedCourse = courseService.create(course);
         var courseId = savedCourse.getId();
+
         var studentId = UUID.randomUUID();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isNotFound());
 
-        courseService.deleteById(courseId);
+        cleanUp(courseId, studentId);
     }
 
     @Test
-    @WithMockUser
     void enrollStudent_givenCourseIdAndStudent_shouldThrowEntityNotFoundExceptionAndReturn404() throws Exception {
         // given
         var courseId = UUID.randomUUID();
+
         var student = buildStudentWithCoins(BigDecimal.ZERO);
         var savedStudent = studentService.create(student);
         var studentId = savedStudent.getId();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isNotFound());
 
-        studentService.deleteById(studentId);
+        cleanUp(courseId, studentId);
     }
 
     @Test
-    @WithMockUser
     void unenrollStudent_givenCourseAndStudent_shouldSuccessfullyUnenrollStudentAndReturn204() throws Exception {
         // given
         var course = buildCourse();
@@ -172,10 +163,12 @@ class CourseEnrollmentControllerIT {
         var student = buildStudentWithCoins(course.getPrice());
         var savedStudent = studentService.create(student);
         var studentId = savedStudent.getId();
+
         var pageable = PageRequest.of(0, 10);
 
         // when
-        mockMvc.perform(delete(ENROLLMENT_URL, courseId, studentId))
+        var request = buildUnenrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isNoContent());
 
         // then
@@ -186,13 +179,15 @@ class CourseEnrollmentControllerIT {
     }
 
     @Test
+    @WithAnonymousUser
     void unenrollStudent_givenCourseIdAndStudentId_shouldReturn401() throws Exception {
         // given
         var courseId = UUID.randomUUID();
         var studentId = UUID.randomUUID();
 
         // when, then
-        mockMvc.perform(delete(ENROLLMENT_URL, courseId, studentId))
+        var request = buildUnenrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isUnauthorized());
     }
 
@@ -210,8 +205,11 @@ class CourseEnrollmentControllerIT {
     }
 
     private void cleanUp(UUID courseId, UUID studentId) {
-        courseService.deleteById(courseId);
-        studentService.deleteById(studentId);
+        if (nonNull(courseId)) {
+            courseService.deleteById(courseId);
+        }
+        if (nonNull(studentId)) {
+            studentService.deleteById(studentId);
+        }
     }
-
 }
